requirements.txt
# Core ML Libraries
tensorflow==2.15.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pillow==10.0.0

# Data Acquisition
kagglehub==0.2.5

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# UI Framework
streamlit==1.28.1
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Load Testing
locust==2.17.0

# AWS SDK
boto3==1.29.7
botocore==1.32.7

# Monitoring
psutil==5.9.6
prometheus-client==0.19.0

# Image Processing
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# Progress bars
tqdm==4.66.1

# Jupyter
jupyter==1.0.0
ipykernel==6.26.0
ipywidgets==8.1.1
